feat: simplify build
Instead of 'ncu' I can just the the built-in 'npm outdated' command, so I can remove that
dependency. As I'm making the switch from Travis to Github Actions with the
'conventional-changelog-action' I will just rely on the actual version of the package i.e. the
version as specified in 'package.json', Using commitizen to ensure I'm following semVer.